Force Ansible to use dynamic includes
Ansible 2.1.1 introduces a regression in the way conditional
includes are handled which results in every task in the
included file being evaluated even if the condition for the
include is not met. This extends the run time significantly
for a deployment.

This patch forces all conditional includes to be dynamic.

Change-Id: I27afcb8e650dc296e06d2bfaa9e084eb406fa4ef
Related-Bug: https://github.com/ansible/ansible/issues/17687