Use package module to install distro packages
Consolidate distro package install tasks into a
single task using the package module and pass
the package list into the name instead of using
a with_items loop.

The minimum Ansible version is raised to 2.2 due to a
known bug [1] in Ansible's apt module which does not
update the cache properly if the cache update and the
install are combined in a single task.

[1] https://github.com/ansible/ansible-modules-core/issues/1497

Also:
- remove unused travis configuration file
- move release note to the proper location
- add license boilerplate to meta file

Change-Id: Ibc2d96a48f156f19235df8b7a25d59d433c3aad5