Allow role defaults to determine the branch tested
Currently there's an override in place to make the tests
use the master branch. This becomes a pain when master
becomes a stable branch because this override is out of
the norm, so it gets forgotten.

The override is not necessary - we can just use the
values from the role defaults instead.

Change-Id: I1f9540951d269545f8466989ef4ff4bfe939d35b