Allow ability to configure number of processes and threads
Users can configure the number of worker threads however when it's
not specified the calculated number of workers can get too large on
hosts with a large number of CPUs.

This also adds the setting of threads and processes to the UWSGI
configuration.

Change-Id: I003ab426488966cce46bd6fd297c79ada13c9668
Closes-Bug: #1745631
(cherry picked from commit 4b0da513d06ebffad24490b9b08eb2033ee6a181)