Only replace python when re-initializing the venv
We reinitialize the venv to ensure that the right version of
python is in the venv, but we do not want virtualenv to also
replace pip, setuptools and wheel so we tell it not to. If we
don't do this then virtualenv will install the latest available
version, which is not what we want.

Change-Id: Ib57a7bc2a88a152e5f60e083c676fb1ce07deb18
Partial-Bug: #1764470
(cherry picked from commit b81c76cf5d27aedf598c038a7a6a99006459d478)